Moves the woff base64 fonts to a woff folder
