+++ src/fonts/index.ts
@@ -1,19 +1,19 @@
 import { FontData } from '../utils/setFont';
-import { helvetica } from './helvetica';
-import { arial } from './arial';
-import { verdana } from './verdana';
-import { tahoma } from './tahoma';
-import { trebuchet_ms } from './trebuchet_ms';
-import { delius } from './delius';
-import { open_sans } from './open_sans';
-import { times_new_roman } from './times_new_roman';
-import { georgia } from './georgia';
-import { garamond } from './garamond';
-import { libre_baskerville } from './libre_baskerville';
-import { courier_new } from './courier_new';
-import { merienda } from './merienda';
-import { sofia } from './sofia';
-import { medieval_sharp } from './medieval_sharp';
+import { helvetica } from './woff/helvetica';
+import { arial } from './woff/arial';
+import { verdana } from './woff/verdana';
+import { tahoma } from './woff/tahoma';
+import { trebuchet_ms } from './woff/trebuchet_ms';
+import { delius } from './woff/delius';
+import { open_sans } from './woff/open_sans';
+import { times_new_roman } from './woff/times_new_roman';
+import { georgia } from './woff/georgia';
+import { garamond } from './woff/garamond';
+import { libre_baskerville } from './woff/libre_baskerville';
+import { courier_new } from './woff/courier_new';
+import { merienda } from './woff/merienda';
+import { sofia } from './woff/sofia';
+import { medieval_sharp } from './woff/medieval_sharp';
 
 const fontFamilies: { [key: string]: FontData } = {
   // San Serif fonts (fallback: san-serif)
